requirements.txt
flask==3.0.3
flask-cors==4.0.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
